simplify sdk and fix skipped tests
