+++ packages/stable/src/__tests__/api/datapoints.int.spec.ts
@@ -4,6 +4,7 @@ import { afterAll, beforeAll, describe, expect, test } from 'vitest';
 import type CogniteClient from '../../cogniteClient';
 import type {
   DatapointAggregate,
+  Datapoints,
   DoubleDatapoint,
   NodeWrite,
   Timeseries,
@@ -167,7 +168,7 @@ describe('Datapoints integration test', () => {
       {
         timestamp: statusTimestamp,
         value: 42,
-        status: { symbol: 'Uncertain' },
+        status: { code: 1073741824, symbol: 'Uncertain' },
       },
     ];
     await client.datapoints.insert([
@@ -361,101 +362,348 @@ describe('Datapoints integration test with chunking', () => {
   });
 });
 
-describe.skip('Datapoints integration test for monthly granularity', () => {
+describe('Datapoints integration test for monthly granularity', () => {
   let client: CogniteClient;
   let timeserie: Timeseries;
   let timeserie2: Timeseries;
+
+  const utc = (y: number, m: number, d: number) => new Date(Date.UTC(y, m, d));
+
   const datapoints = [
-    // Create two data points in October 2022
-    {
-      timestamp: new Date(2022, 9, 1),
-      value: 10,
-    },
-    {
-      timestamp: new Date(2022, 9, 2),
-      value: 20,
-    },
-    // Create two data points in November 2022
-    {
-      timestamp: new Date(2022, 10, 1),
-      value: 30,
-    },
-    {
-      timestamp: new Date(2022, 10, 2),
-      value: 40,
-    },
-    // Create two data points in December 2022
-    {
-      timestamp: new Date(2022, 11, 1),
-      value: 50,
-    },
-    {
-      timestamp: new Date(2022, 11, 2),
-      value: 60,
-    },
-    // Create two data points in January 2023
-    {
-      timestamp: new Date(2023, 0, 1),
-      value: 70,
-    },
-    {
-      timestamp: new Date(2023, 0, 2),
-      value: 80,
-    },
-    // Create a missing month in between, populated with data points in March 2023
-    {
-      timestamp: new Date(2023, 2, 1),
-      value: 90,
-    },
-    {
-      timestamp: new Date(2023, 2, 2),
-      value: 100,
-    },
+    { timestamp: utc(2022, 9, 1), value: 10 },
+    { timestamp: utc(2022, 9, 2), value: 20 },
+    { timestamp: utc(2022, 10, 1), value: 30 },
+    { timestamp: utc(2022, 10, 2), value: 40 },
+    { timestamp: utc(2022, 11, 1), value: 50 },
+    { timestamp: utc(2022, 11, 2), value: 60 },
+    { timestamp: utc(2023, 0, 1), value: 70 },
+    { timestamp: utc(2023, 0, 2), value: 80 },
+    { timestamp: utc(2023, 2, 1), value: 90 },
+    { timestamp: utc(2023, 2, 2), value: 100 },
   ];
   const datapoints2 = [
-    // Create two data points in October 2022
+    { timestamp: utc(2022, 9, 1), value: 0 },
+    { timestamp: utc(2022, 9, 2), value: 10 },
+    { timestamp: utc(2022, 10, 1), value: 20 },
+    { timestamp: utc(2022, 10, 2), value: 30 },
+    { timestamp: utc(2022, 11, 1), value: 40 },
+    { timestamp: utc(2022, 11, 2), value: 50 },
+    { timestamp: utc(2023, 1, 1), value: 60 },
+    { timestamp: utc(2023, 1, 2), value: 70 },
+    { timestamp: utc(2023, 2, 1), value: 80 },
+    { timestamp: utc(2023, 2, 2), value: 90 },
+  ];
+
+  beforeAll(async () => {
+    client = setupLoggedInClient();
+    [timeserie] = await client.timeseries.create([{ name: 'tmp' }]);
+    await client.datapoints.insert([
+      { id: timeserie.id, datapoints: datapoints },
+    ]);
+
+    [timeserie2] = await client.timeseries.create([{ name: 'tmp2' }]);
+    await client.datapoints.insert([
+      { id: timeserie2.id, datapoints: datapoints2 },
+    ]);
+  });
+  afterAll(async () => {
+    await client.timeseries.delete([{ id: timeserie.id }]);
+    await client.timeseries.delete([{ id: timeserie2.id }]);
+  });
+
+  test('retrieve monthly granularity for two consecutive months in same year', async () => {
+    const response = await client.datapoints.retrieveDatapointMonthlyAggregates(
+      {
+        items: [{ id: timeserie.id }, { id: timeserie2.id }],
+        start: Date.UTC(2022, 9, 1),
+        end: Date.UTC(2022, 10, 30),
+        aggregates: ['sum'],
+      }
+    );
+
+    expect(response[0].datapoints.length).toBe(2);
+    expect((response[0].datapoints[0] as DatapointAggregate).sum).toBe(30);
+    expect((response[0].datapoints[1] as DatapointAggregate).sum).toBe(70);
+    expect((response[0].datapoints[0] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 9, 1)
+    );
+    expect((response[0].datapoints[1] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 10, 1)
+    );
+
+    expect(response[1].datapoints.length).toBe(2);
+    expect((response[1].datapoints[0] as DatapointAggregate).sum).toBe(10);
+    expect((response[1].datapoints[1] as DatapointAggregate).sum).toBe(50);
+    expect((response[1].datapoints[0] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 9, 1)
+    );
+    expect((response[1].datapoints[1] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 10, 1)
+    );
+  });
+
+  test('retrieve monthly granularity for two consecutive months in different years', async () => {
+    const response = await client.datapoints.retrieveDatapointMonthlyAggregates(
+      {
+        items: [{ id: timeserie.id }],
+        start: Date.UTC(2022, 11, 1),
+        end: Date.UTC(2023, 0, 15),
+        aggregates: ['sum'],
+      }
+    );
+    expect(response[0].datapoints.length).toBe(2);
+    expect((response[0].datapoints[0] as DatapointAggregate).sum).toBe(110);
+    expect((response[0].datapoints[1] as DatapointAggregate).sum).toBe(150);
+    expect((response[0].datapoints[0] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 11, 1)
+    );
+    expect((response[0].datapoints[1] as DatapointAggregate).timestamp).toEqual(
+      utc(2023, 0, 1)
+    );
+  });
+
+  test('retrieve monthly granularity for two non-consecutive months in same year', async () => {
+    const response = await client.datapoints.retrieveDatapointMonthlyAggregates(
+      {
+        items: [{ id: timeserie.id }],
+        start: Date.UTC(2022, 9, 1),
+        end: Date.UTC(2022, 11, 30),
+        aggregates: ['sum'],
+      }
+    );
+    expect(response[0].datapoints.length).toBe(3);
+    expect((response[0].datapoints[0] as DatapointAggregate).sum).toBe(30);
+    expect((response[0].datapoints[1] as DatapointAggregate).sum).toBe(70);
+    expect((response[0].datapoints[2] as DatapointAggregate).sum).toBe(110);
+    expect((response[0].datapoints[0] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 9, 1)
+    );
+    expect((response[0].datapoints[1] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 10, 1)
+    );
+    expect((response[0].datapoints[2] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 11, 1)
+    );
+  });
+
+  test('retrieve monthly granularity when there is a data gap between months', async () => {
+    const response = await client.datapoints.retrieveDatapointMonthlyAggregates(
+      {
+        items: [{ id: timeserie.id }, { id: timeserie2.id }],
+        start: Date.UTC(2022, 9, 1),
+        end: Date.UTC(2023, 2, 15),
+        aggregates: ['sum'],
+      }
+    );
+    expect(response[0].datapoints.length).toBe(5);
+    expect((response[0].datapoints[0] as DatapointAggregate).sum).toBe(30);
+    expect((response[0].datapoints[1] as DatapointAggregate).sum).toBe(70);
+    expect((response[0].datapoints[2] as DatapointAggregate).sum).toBe(110);
+    expect((response[0].datapoints[3] as DatapointAggregate).sum).toBe(150);
+    expect((response[0].datapoints[4] as DatapointAggregate).sum).toBe(190);
+    expect((response[0].datapoints[0] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 9, 1)
+    );
+    expect((response[0].datapoints[1] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 10, 1)
+    );
+    expect((response[0].datapoints[2] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 11, 1)
+    );
+    expect((response[0].datapoints[3] as DatapointAggregate).timestamp).toEqual(
+      utc(2023, 0, 1)
+    );
+    expect((response[0].datapoints[4] as DatapointAggregate).timestamp).toEqual(
+      utc(2023, 2, 1)
+    );
+
+    expect(response[1].datapoints.length).toBe(5);
+    expect((response[1].datapoints[0] as DatapointAggregate).sum).toBe(10);
+    expect((response[1].datapoints[1] as DatapointAggregate).sum).toBe(50);
+    expect((response[1].datapoints[2] as DatapointAggregate).sum).toBe(90);
+    expect((response[1].datapoints[3] as DatapointAggregate).sum).toBe(130);
+    expect((response[1].datapoints[4] as DatapointAggregate).sum).toBe(170);
+    expect((response[1].datapoints[0] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 9, 1)
+    );
+    expect((response[1].datapoints[1] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 10, 1)
+    );
+    expect((response[1].datapoints[2] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 11, 1)
+    );
+    expect((response[1].datapoints[3] as DatapointAggregate).timestamp).toEqual(
+      utc(2023, 1, 1)
+    );
+    expect((response[1].datapoints[4] as DatapointAggregate).timestamp).toEqual(
+      utc(2023, 2, 1)
+    );
+  });
+
+  test('retrieve monthly granularity for a year when there is missing data for some months', async () => {
+    const response = await client.datapoints.retrieveDatapointMonthlyAggregates(
+      {
+        items: [{ id: timeserie.id }],
+        start: Date.UTC(2022, 0, 1),
+        end: Date.UTC(2022, 11, 15),
+        aggregates: ['sum'],
+      }
+    );
+    expect(response[0].datapoints.length).toBe(3);
+    expect((response[0].datapoints[0] as DatapointAggregate).sum).toBe(30);
+    expect((response[0].datapoints[1] as DatapointAggregate).sum).toBe(70);
+    expect((response[0].datapoints[2] as DatapointAggregate).sum).toBe(110);
+    expect((response[0].datapoints[0] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 9, 1)
+    );
+    expect((response[0].datapoints[1] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 10, 1)
+    );
+    expect((response[0].datapoints[2] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 11, 1)
+    );
+  });
+
+  test('retrieve monthly average granularity when there is a data gap between months', async () => {
+    const response = await client.datapoints.retrieveDatapointMonthlyAggregates(
+      {
+        items: [{ id: timeserie.id }],
+        start: Date.UTC(2022, 9, 1),
+        end: Date.UTC(2023, 2, 15),
+        aggregates: ['average'],
+      }
+    );
+    expect(response[0].datapoints.length).toBe(5);
+    expect((response[0].datapoints[0] as DatapointAggregate).average).toBe(15);
+    expect((response[0].datapoints[1] as DatapointAggregate).average).toBe(35);
+    expect((response[0].datapoints[2] as DatapointAggregate).average).toBe(55);
+    expect((response[0].datapoints[3] as DatapointAggregate).average).toBe(75);
+    expect((response[0].datapoints[4] as DatapointAggregate).average).toBe(95);
+    expect((response[0].datapoints[0] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 9, 1)
+    );
+    expect((response[0].datapoints[1] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 10, 1)
+    );
+    expect((response[0].datapoints[2] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 11, 1)
+    );
+    expect((response[0].datapoints[3] as DatapointAggregate).timestamp).toEqual(
+      utc(2023, 0, 1)
+    );
+    expect((response[0].datapoints[4] as DatapointAggregate).timestamp).toEqual(
+      utc(2023, 2, 1)
+    );
+  });
+
+  test('retrieve monthly granularity with UTC start/end', async () => {
+    const response = await client.datapoints.retrieveDatapointMonthlyAggregates(
+      {
+        items: [{ id: timeserie.id }],
+        start: Date.UTC(2022, 0, 1),
+        end: Date.UTC(2022, 11, 12),
+        aggregates: ['sum'],
+      }
+    );
+    expect(response[0].datapoints.length).toBe(3);
+    expect((response[0].datapoints[0] as DatapointAggregate).sum).toBe(30);
+    expect((response[0].datapoints[1] as DatapointAggregate).sum).toBe(70);
+    expect((response[0].datapoints[2] as DatapointAggregate).sum).toBe(110);
+    expect((response[0].datapoints[0] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 9, 1)
+    );
+    expect((response[0].datapoints[1] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 10, 1)
+    );
+    expect((response[0].datapoints[2] as DatapointAggregate).timestamp).toEqual(
+      utc(2022, 11, 1)
+    );
+  });
+
+  test('retrieve monthly granularity with empty result set', async () => {
+    const response = await client.datapoints.retrieveDatapointMonthlyAggregates(
+      {
+        items: [{ id: timeserie.id }, { id: timeserie2.id }],
+        start: Date.UTC(2020, 9, 1),
+        end: Date.UTC(2020, 10, 30),
+        aggregates: ['sum'],
+      }
+    );
+
+    expect(response[0].datapoints.length).toBe(0);
+    expect(response[1].datapoints.length).toBe(0);
+  });
+});
+
+const Status = {
+  good: {
+    code: 0,
+    symbol: 'Good',
+  },
+  uncertain: {
+    code: 1073741824,
+    symbol: 'Uncertain',
+  },
+  bad: {
+    code: 2147483648,
+    symbol: 'Bad',
+  },
+};
+
+describe('Datapoints integration test for quality indicators', () => {
+  let client: CogniteClient;
+  let timeserie: Timeseries;
+
+  const datapoints = [
     {
-      timestamp: new Date(2022, 9, 1),
-      value: 0,
-    },
-    {
-      timestamp: new Date(2022, 9, 2),
+      timestamp: new Date('2022-10-01T00:00:00Z'),
       value: 10,
+      status: Status.good,
     },
-    // Create two data points in November 2022
     {
-      timestamp: new Date(2022, 10, 1),
+      timestamp: new Date('2022-10-02T00:00:00Z'),
       value: 20,
+      status: Status.good,
     },
     {
-      timestamp: new Date(2022, 10, 2),
+      timestamp: new Date('2022-11-01T00:00:00Z'),
       value: 30,
+      status: Status.uncertain,
     },
-    // Create two data points in December 2022
     {
-      timestamp: new Date(2022, 11, 1),
+      timestamp: new Date('2022-11-02T00:00:00Z'),
       value: 40,
+      status: Status.uncertain,
     },
     {
-      timestamp: new Date(2022, 11, 2),
+      timestamp: new Date('2022-12-01T00:00:00Z'),
       value: 50,
+      status: Status.bad,
     },
-    // Create a missing month in between, populated with data points in Feb 2023
     {
-      timestamp: new Date(2023, 1, 1),
+      timestamp: new Date('2022-12-02T00:00:00Z'),
       value: 60,
+      status: Status.bad,
     },
     {
-      timestamp: new Date(2023, 1, 2),
+      timestamp: new Date('2023-01-01T00:00:00Z'),
       value: 70,
+      status: Status.good,
     },
     {
-      timestamp: new Date(2023, 2, 1),
+      timestamp: new Date('2023-01-02T00:00:00Z'),
       value: 80,
+      status: Status.bad,
     },
     {
-      timestamp: new Date(2023, 2, 2),
+      timestamp: new Date('2023-03-01T00:00:00Z'),
       value: 90,
+      status: Status.good,
+    },
+    {
+      timestamp: new Date('2023-03-02T00:00:00Z'),
+      value: 100,
+      status: Status.uncertain,
     },
   ];
 
@@ -468,285 +716,65 @@ describe.skip('Datapoints integration test for monthly granularity', () => {
         datapoints: datapoints,
       },
     ]);
-
-    [timeserie2] = await client.timeseries.create([{ name: 'tmp2' }]);
-    await client.datapoints.insert([
-      {
-        id: timeserie2.id,
-        datapoints: datapoints2,
-      },
-    ]);
   });
   afterAll(async () => {
     await client.timeseries.delete([{ id: timeserie.id }]);
-    await client.timeseries.delete([{ id: timeserie2.id }]);
   });
 
-  test('retrieve monthly granularity for two consecutive months in same year', async () => {
+  test('retrieve aggregates with correct data point status counts', async () => {
     const response = await client.datapoints.retrieveDatapointMonthlyAggregates(
       {
-        items: [{ id: timeserie.id }, { id: timeserie2.id }],
-        start: new Date(2022, 9, 1),
-        end: new Date(2022, 10, 30),
-        aggregates: ['sum'],
+        items: [{ id: timeserie.id, includeStatus: true }],
+        start: new Date(2020, 0, 1),
+        end: new Date(2023, 3, 1),
+        aggregates: ['countGood', 'countUncertain', 'countBad'],
       }
     );
 
-    expect(response[0].datapoints.length).toBe(2);
-    // Check that the response contains the correct number of data points
-    expect((response[0].datapoints[0] as DatapointAggregate).sum).toBe(30);
-    expect((response[0].datapoints[1] as DatapointAggregate).sum).toBe(70);
-    // Check timestamps
-    expect((response[0].datapoints[0] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 9, 1)
-    );
-    expect((response[0].datapoints[1] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 10, 1)
-    );
+    expect(response[0].datapoints[0].countGood).toBe(2);
+    expect(response[0].datapoints[0].countUncertain).toBe(0);
+    expect(response[0].datapoints[0].countBad).toBe(0);
 
-    // check that there is two timeseries in the response
-    expect(response[1].datapoints.length).toBe(2);
-    // Check that the response contains the correct number of data points
-    expect((response[1].datapoints[0] as DatapointAggregate).sum).toBe(10);
-    expect((response[1].datapoints[1] as DatapointAggregate).sum).toBe(50);
-    // Check timestamps
-    expect((response[1].datapoints[0] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 9, 1)
-    );
-    expect((response[1].datapoints[1] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 10, 1)
-    );
+    expect(response[0].datapoints[1].countGood).toBe(0);
+    expect(response[0].datapoints[1].countUncertain).toBe(2);
+    expect(response[0].datapoints[1].countBad).toBe(0);
+
+    expect(response[0].datapoints[2].countGood).toBe(0);
+    expect(response[0].datapoints[2].countUncertain).toBe(0);
+    expect(response[0].datapoints[2].countBad).toBe(2);
+
+    expect(response[0].datapoints[3].countGood).toBe(1);
+    expect(response[0].datapoints[3].countUncertain).toBe(0);
+    expect(response[0].datapoints[3].countBad).toBe(1);
+
+    expect(response[0].datapoints[4].countGood).toBe(1);
+    expect(response[0].datapoints[4].countUncertain).toBe(1);
+    expect(response[0].datapoints[4].countBad).toBe(0);
   });
 
-  test('retrieve monthly granularity for two consecutive months in different years', async () => {
-    const response = await client.datapoints.retrieveDatapointMonthlyAggregates(
-      {
-        items: [{ id: timeserie.id }],
-        start: new Date(2022, 11, 1),
-        end: new Date(2023, 0, 15),
-        aggregates: ['sum'],
-      }
-    );
-    expect(response[0].datapoints.length).toBe(2);
-    // Check that the response contains the correct number of data points
-    expect((response[0].datapoints[0] as DatapointAggregate).sum).toBe(110);
-    expect((response[0].datapoints[1] as DatapointAggregate).sum).toBe(150);
-    expect((response[0].datapoints[0] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 11, 1)
-    );
-    expect((response[0].datapoints[1] as DatapointAggregate).timestamp).toEqual(
-      new Date(2023, 0, 1)
-    );
-  });
+  test('retrieve datapoints with status code', async () => {
+    const response = (await client.datapoints.retrieve({
+      items: [
+        {
+          id: timeserie.id,
+          start: new Date('2022-10-01T00:00:00Z'),
+          end: new Date('2023-04-01T00:00:00Z'),
+          includeStatus: true,
+          ignoreBadDataPoints: false,
+          treatUncertainAsBad: false,
+        },
+      ],
+    })) as Datapoints[];
 
-  test('retrieve monthly granularity for two non-consecutive months in same year', async () => {
-    const response = await client.datapoints.retrieveDatapointMonthlyAggregates(
-      {
-        items: [{ id: timeserie.id }],
-        start: new Date(2022, 9, 1),
-        end: new Date(2022, 11, 30),
-        aggregates: ['sum'],
-      }
-    );
-    expect(response[0].datapoints.length).toBe(3);
-    // Check that the response contains the correct number of data points
-    expect((response[0].datapoints[0] as DatapointAggregate).sum).toBe(30);
-    expect((response[0].datapoints[1] as DatapointAggregate).sum).toBe(70);
-    expect((response[0].datapoints[2] as DatapointAggregate).sum).toBe(110);
-    expect((response[0].datapoints[0] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 9, 1)
-    );
-    expect((response[0].datapoints[1] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 10, 1)
-    );
-    expect((response[0].datapoints[2] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 11, 1)
-    );
-  });
-
-  test('retrieve monthly granularity when there is a data gap between months', async () => {
-    const response = await client.datapoints.retrieveDatapointMonthlyAggregates(
-      {
-        items: [{ id: timeserie.id }, { id: timeserie2.id }],
-        start: new Date(2022, 9, 1),
-        end: new Date(2023, 2, 15),
-        aggregates: ['sum'],
-      }
-    );
-    expect(response[0].datapoints.length).toBe(5);
-
-    // Check that the response contains the correct number of data points
-    expect((response[0].datapoints[0] as DatapointAggregate).sum).toBe(30); // October 2022
-    expect((response[0].datapoints[1] as DatapointAggregate).sum).toBe(70); // November 2022
-    expect((response[0].datapoints[2] as DatapointAggregate).sum).toBe(110); // December 2022
-    expect((response[0].datapoints[3] as DatapointAggregate).sum).toBe(150); // January 2023
-    expect((response[0].datapoints[4] as DatapointAggregate).sum).toBe(190); // March 2023
-    // Check timestamps
-    expect((response[0].datapoints[0] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 9, 1)
-    );
-    expect((response[0].datapoints[1] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 10, 1)
-    );
-    expect((response[0].datapoints[2] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 11, 1)
-    );
-    expect((response[0].datapoints[3] as DatapointAggregate).timestamp).toEqual(
-      new Date(2023, 0, 1)
-    );
-    expect((response[0].datapoints[4] as DatapointAggregate).timestamp).toEqual(
-      new Date(2023, 2, 1)
-    );
-
-    expect(response[1].datapoints.length).toBe(5);
-
-    // Check that the response contains the correct number of data points
-    expect((response[1].datapoints[0] as DatapointAggregate).sum).toBe(10); // October 2022
-    expect((response[1].datapoints[1] as DatapointAggregate).sum).toBe(50); // November 2022
-    expect((response[1].datapoints[2] as DatapointAggregate).sum).toBe(90); // December 2022
-    expect((response[1].datapoints[3] as DatapointAggregate).sum).toBe(130); // February 2023
-    expect((response[1].datapoints[4] as DatapointAggregate).sum).toBe(170); // March 2023
-    // Check timestamps
-    expect((response[1].datapoints[0] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 9, 1)
-    );
-    expect((response[1].datapoints[1] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 10, 1)
-    );
-    expect((response[1].datapoints[2] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 11, 1)
-    );
-    expect((response[1].datapoints[3] as DatapointAggregate).timestamp).toEqual(
-      new Date(2023, 1, 1)
-    );
-    expect((response[0].datapoints[4] as DatapointAggregate).timestamp).toEqual(
-      new Date(2023, 2, 1)
-    );
-  });
-
-  test('retrieve monthly granularity for a year when there is missing data for some months', async () => {
-    const response = await client.datapoints.retrieveDatapointMonthlyAggregates(
-      {
-        items: [{ id: timeserie.id }],
-        start: Date.parse('2022-01-01T00:00:00Z'),
-        end: Date.parse('2022-12-15T23:59:59Z'),
-        aggregates: ['sum'],
-      }
-    );
-    expect(response[0].datapoints.length).toBe(3);
-
-    // Check that the response contains the correct number of data points
-    expect((response[0].datapoints[0] as DatapointAggregate).sum).toBe(30); // October 2022
-    expect((response[0].datapoints[1] as DatapointAggregate).sum).toBe(70); // November 2022
-    expect((response[0].datapoints[2] as DatapointAggregate).sum).toBe(110); // December 2022
-    // Check timestamps
-    expect((response[0].datapoints[0] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 9, 1)
-    );
-    expect((response[0].datapoints[1] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 10, 1)
-    );
-    expect((response[0].datapoints[2] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 11, 1)
-    );
-  });
-
-  test('retrieve monthly average granularity when there is a data gap between months', async () => {
-    const response = await client.datapoints.retrieveDatapointMonthlyAggregates(
-      {
-        items: [{ id: timeserie.id }],
-        start: new Date(2022, 9, 1),
-        end: new Date(2023, 2, 15),
-        aggregates: ['average'],
-      }
-    );
-    expect(response[0].datapoints.length).toBe(5);
-
-    // Check that the response contains the correct number of data points
-    expect((response[0].datapoints[0] as DatapointAggregate).average).toBe(15);
-    expect((response[0].datapoints[1] as DatapointAggregate).average).toBe(35);
-    expect((response[0].datapoints[2] as DatapointAggregate).average).toBe(55);
-    expect((response[0].datapoints[3] as DatapointAggregate).average).toBe(75);
-    expect((response[0].datapoints[4] as DatapointAggregate).average).toBe(95);
-    // Check timestamps
-    expect((response[0].datapoints[0] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 9, 1)
-    );
-    expect((response[0].datapoints[1] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 10, 1)
-    );
-    expect((response[0].datapoints[2] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 11, 1)
-    );
-    expect((response[0].datapoints[3] as DatapointAggregate).timestamp).toEqual(
-      new Date(2023, 0, 1)
-    );
-    expect((response[0].datapoints[4] as DatapointAggregate).timestamp).toEqual(
-      new Date(2023, 2, 1)
-    );
-  });
-
-  test('retrieve monthly granularity with local time zone in start/end - +0100', async () => {
-    const response = await client.datapoints.retrieveDatapointMonthlyAggregates(
-      {
-        items: [{ id: timeserie.id }],
-        start: Date.parse('2022-01-01T00:00:00+0100'),
-        end: Date.parse('2022-12-12T00:00:00+0100'),
-        aggregates: ['sum'],
-      }
-    );
-    expect(response[0].datapoints.length).toBe(3);
-    // Check that the response contains the correct number of data points
-    expect((response[0].datapoints[0] as DatapointAggregate).sum).toBe(30); // October 2022
-    expect((response[0].datapoints[1] as DatapointAggregate).sum).toBe(70); // November 2022
-    expect((response[0].datapoints[2] as DatapointAggregate).sum).toBe(110); // December 2022
-    // Check timestamps
-    expect((response[0].datapoints[0] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 9, 1)
-    );
-    expect((response[0].datapoints[1] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 10, 1)
-    );
-    expect((response[0].datapoints[2] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 11, 1)
-    );
-  });
-  test('retrieve monthly granularity with local time zone in start/end - -0600', async () => {
-    const response = await client.datapoints.retrieveDatapointMonthlyAggregates(
-      {
-        items: [{ id: timeserie.id }],
-        start: Date.parse('2022-01-01T00:00:00-0600'),
-        end: Date.parse('2022-12-12T00:00:00-0600'),
-        aggregates: ['sum'],
-      }
-    );
-    expect(response[0].datapoints.length).toBe(3);
-    // Check that the response contains the correct number of data points
-    expect((response[0].datapoints[0] as DatapointAggregate).sum).toBe(30); // October 2022
-    expect((response[0].datapoints[1] as DatapointAggregate).sum).toBe(70); // November 2022
-    expect((response[0].datapoints[2] as DatapointAggregate).sum).toBe(110); // December 2022
-    // Check timestamps
-    expect((response[0].datapoints[0] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 9, 1)
-    );
-    expect((response[0].datapoints[1] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 10, 1)
-    );
-    expect((response[0].datapoints[2] as DatapointAggregate).timestamp).toEqual(
-      new Date(2022, 11, 1)
-    );
-  });
-  test('retrieve monthly granularity with empty result ste', async () => {
-    const response = await client.datapoints.retrieveDatapointMonthlyAggregates(
-      {
-        items: [{ id: timeserie.id }, { id: timeserie2.id }],
-        start: new Date(2020, 9, 1),
-        end: new Date(2020, 10, 30),
-        aggregates: ['sum'],
-      }
-    );
-
-    expect(response[0].datapoints.length).toBe(0);
-    expect(response[1].datapoints.length).toBe(0);
+    expect(response[0].datapoints[0].status).toBeUndefined();
+    expect(response[0].datapoints[1].status).toBeUndefined();
+    expect(response[0].datapoints[2].status).toEqual(Status.uncertain);
+    expect(response[0].datapoints[3].status).toEqual(Status.uncertain);
+    expect(response[0].datapoints[4].status).toEqual(Status.bad);
+    expect(response[0].datapoints[5].status).toEqual(Status.bad);
+    expect(response[0].datapoints[6].status).toBeUndefined();
+    expect(response[0].datapoints[7].status).toEqual(Status.bad);
+    expect(response[0].datapoints[8].status).toBeUndefined();
+    expect(response[0].datapoints[9].status).toEqual(Status.uncertain);
   });
 });
